move GenericTraversal to new module
