+++ core/src/lib.rs
@@ -32,6 +32,7 @@ pub mod optimistic_confirmation_verifier;
 pub mod outstanding_requests;
 pub mod packet_hasher;
 pub mod progress_map;
+pub mod repair_generic_traversal;
 pub mod repair_response;
 pub mod repair_service;
 pub mod repair_weight;
